+++ include/output.h
@@ -61,8 +61,7 @@ class Output {
 
   void commit();
 
-  void render(const std::vector<std::shared_ptr<View>>& views,
-    const std::vector<std::shared_ptr<View>>& overlays) const;
+  void render(const std::vector<std::shared_ptr<View>>& views) const;
 
   void take_damage(const View *view);
   void take_whole_damage();
