+++ include/shell.h
@@ -3,7 +3,6 @@
 
 #include <memory>
 #include <thread>
-#include <vector>
 
 namespace lumin {
 
@@ -12,18 +11,20 @@ class Switcher;
 
 class Shell {
  public:
+  ~Shell();
   Shell(Server *server);
+
  public:
   void run();
   void switch_app();
   void switch_app_reverse();
   void cancel_activity();
+
  private:
   void init();
   static void thread(void *data);
+
   std::thread ui_thread_;
-  int argc_;
-  char **argv_;
   std::unique_ptr<Switcher> switcher_;
 };
