+++ include/shell/switcher/view.h
@@ -1,14 +1,12 @@
-#ifndef SWITCHER_H_
-#define SWITCHER_H_
+#ifndef SWITCHER_VIEW_H_
+#define SWITCHER_VIEW_H_
 
 #include <gtk/gtk.h>
 
-#include <memory>
 #include <vector>
 
 namespace lumin {
 
-class Server;
 class View;
 
 class SwitcherView {
@@ -26,20 +24,6 @@ class SwitcherView {
   GtkWidget *box_;
 };
 
-class Switcher {
- public:
-  Switcher(Server *server);
- public:
-  void init();
-  void show();
-  void previous();
-  void hide();
- private:
-  std::unique_ptr<SwitcherView> view_;
-  Server *server_;
-  int selected_index_;
-};
-
 }  // namespace lumin
 
-#endif  // SWITCHER_H_
+#endif  // SWITCHER_VIEW_H_
