+++ src/key_binding.cpp
@@ -33,7 +33,6 @@ int KeyBinding::mods() const {
 }
 
 bool KeyBinding::matches(int modifiers, unsigned int sym, wlr_key_state key_state) {
-  spdlog::debug("{} {} {} {}", mods(), modifiers, sym, key);
   bool match = (mods() == modifiers && sym == key && state == key_state);
   return match;
 }
