+++ src/shell.cpp
@@ -3,10 +3,14 @@
 #include <gtk/gtk.h>
 #include <spdlog/spdlog.h>
 
-#include "switcher.h"
-
+#include "shell/switcher/switcher.h"
+#include "shell/switcher/view.h"
 
 namespace lumin {
+
+Shell::~Shell() {
+
+}
 
 Shell::Shell(Server *server) {
   switcher_ = std::make_unique<Switcher>(server);
@@ -14,12 +18,12 @@ Shell::Shell(Server *server) {
 
 void Shell::thread(void *data) {
   Shell *shell = static_cast<Shell*>(data);
+  gtk_init(0, NULL);
   shell->init();
   gtk_main();
 }
 
 void Shell::init() {
-  gtk_init(0, NULL);
   switcher_->init();
 }
 
@@ -29,7 +33,7 @@ void Shell::run() {
 
 int switch_app_gui_thread(void *data) {
   auto *switcher = static_cast<Switcher*>(data);
-  switcher->show();
+  switcher->next();
   return 0;
 }
